Complete at a basic level. No error checking as of yet and only tested image caching so far.
Image caching appears to work, more formal tests yet to be carried out.
The index.php file gives a general overview of how it would be used, you set the variables first (some could be defined throughout page of course) and then create a new caching object for whatever API call you want to make and then use getAPIResults() to get the data back. In the case of an image it is the URI of the image file, otherwise it is the raw API response.

Still todo:
 - Create a README file
 - Introduce Error checking
 - Use parameters (or some hash thereof) to make unique caches for different parameter combinations to same API function
 - More testing
 - etc etc...