A lot of restructuring and improvements
A lot of restructuring has gone into this version, as livestats API function call does not need to be cached and so simply returns direct API results.
The first 20 characters of the MD5 hash of the parameter string is used as a unique identifier for the cache files.
cURL no longer suffers from an issue when used on an https protocol.